add support for sequel
